src/back-end/AgroHelp/requirements.txt
adrf==0.1.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
Deprecated==1.2.13
Django==4.2
django-cors-headers==3.14.0
django-filter==23.1
django-leaflet==0.28.3
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
drf-yasg==1.21.5
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.0
pika==1.3.1
Pillow==9.4.0
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
redis==4.5.4
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==6.0
